+++ mltu/__init__.py
@@ -1,1 +1,3 @@
-__version__ = "1.0.2"
+__version__ = "1.0.3"
+
+from .annotations.image import Image